move examples to examples folder; add doc explaining updated framework
