+++ tests/examples/generation_configs/custom_class/run.yaml
