+++ tests/examples/generation_configs/custom_script/test_python_script.yaml
